# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: OpenAPI documentation for Bank Cards service
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /users/{id}:
    put:
      tags:
      - Users
      summary: Обновить пользователя по ID
      description: Обновляет данные пользователя с указанным ID
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "200":
          description: Пользователь успешно обновлен
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Ошибка валидации или при обновлении
        "403":
          description: Ошибка авторизации
        "404":
          description: Пользователь не найден
        "500":
          description: Ошибка базы данных или роли
      security:
      - bearerAuth: []
    delete:
      tags:
      - Users
      summary: Удалить пользователя по ID
      description: Удаляет пользователя с указанным ID
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Пользователь успешно удалён
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DeleteUserResponse"
        "403":
          description: Ошибка авторизации
        "404":
          description: Пользователь не найден
        "500":
          description: Ошибка базы данных
      security:
      - bearerAuth: []
...
  /activities/block-requests/{requestId}:
    put:
      tags:
      - User Activities
      summary: Завершить запрос на блокировку карты
      description: Отмечает запрос на блокировку карты как завершенный
      operationId: completeBlockRequest
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Запрос успешно завершен
        "400":
          description: Ошибка в запросе
        "403":
          description: Ошибка авторизации
        "404":
          description: Запрос не найден
        "500":
          description: Внутренняя ошибка сервиса
      security:
      - bearerAuth: []
  /users:
    get:
      tags:
      - Users
      summary: Получить список пользователей
      description: Возвращает список пользователей с пагинацией
      operationId: getUsers
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: Список пользователей
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Page"
        "403":
          description: Ошибка авторизации
        "500":
          description: Ошибка базы данных
      security:
      - bearerAuth: []
    post:
      tags:
      - Users
      summary: Создать нового пользователя
      description: Создает пользователя с заданными данными
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          description: Пользователь успешно создан
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Ошибка валидации или при обновлении
        "403":
          description: Ошибка авторизации
        "500":
          description: Ошибка базы данных или роли
      security:
      - bearerAuth: []
...
  /activities/transfer:
    post:
      tags:
      - User Activities
      summary: Перевод денег между картами
      description: Позволяет перевести деньги с одной карты пользователя на другую
        карту
      operationId: transferMoney
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
        required: true
      responses:
        "200":
          description: Перевод успешно выполнен
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardTransfer"
        "400":
          description: Ошибка в запросе или параметрах
        "403":
          description: Ошибка авторизации
        "404":
          description: Карта не найдена
        "500":
          description: Ошибка базы данных или сервиса
      security:
      - bearerAuth: []
  /activities/block-request/{cardId}:
    post:
      tags:
      - User Activities
      summary: Запрос на блокировку карты
      description: Создает запрос на блокировку карты пользователя
      operationId: requestBlockCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "201":
          description: Запрос на блокировку успешно создан
          content:
            '*/*':
              schema:
                type: integer
                format: int64
        "400":
          description: Ошибка в параметрах запроса
        "403":
          description: Ошибка авторизации
        "404":
          description: Карта не найдена
        "500":
          description: Внутренняя ошибка сервиса
      security:
      - bearerAuth: []
...
  /activities/block-requests:
    get:
      tags:
      - User Activities
      summary: Получить активные запросы на блокировку карт
      description: Возвращает страницу с активными запросами на блокировку карт
      operationId: getActiveBlockRequests
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: Список активных запросов получен
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Page"
        "403":
          description: Ошибка авторизации
        "500":
          description: Ошибка базы данных
      security:
      - bearerAuth: []
components:
  schemas:
    UserResponse:
      type: object
      description: Ответ с данными пользователя
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор пользователя
          example: 123
        login:
          type: string
          description: Логин пользователя
          example: user123
        role:
          type: string
          description: Роль пользователя
          example: ADMIN
    UpdateUserRequest:
      type: object
      description: Запрос для обновления пользователя
      properties:
        login:
          type: string
          description: Логин пользователя
          example: user123
        password:
          type: string
          description: Пароль пользователя
          example: secretPassword
        roleName:
          type: string
          description: Роль пользователя
          example: ADMIN
...
    UserRequest:
      type: object
      description: Запрос на создание пользователя
      properties:
        login:
          type: string
          description: Логин пользователя
          example: user123
          minLength: 1
        password:
          type: string
          description: Пароль пользователя
          example: secretPassword
          minLength: 1
        roleName:
          type: string
          description: Название роли пользователя
          example: USER
          minLength: 1
      required:
      - login
      - password
      - roleName
...
    Card:
      type: object
      properties:
        id:
          type: integer
          format: int64
        expiryDate:
          type: string
          format: date
        balance:
          type: number
        status:
          $ref: "#/components/schemas/CardStatus"
        owner:
          $ref: "#/components/schemas/User"
        maskedCardNumber:
          type: string
    CardStatus:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
    CardTransfer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fromCard:
          $ref: "#/components/schemas/Card"
        toCard:
          $ref: "#/components/schemas/Card"
        amount:
          type: number
        transferTime:
          type: string
          format: date-time
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        enabled:
          type: boolean
        accountNonExpired:
          type: boolean
        credentialsNonExpired:
          type: boolean
        username:
          type: string
        accountNonLocked:
          type: boolean
      required:
      - login
    TransferRequest:
      type: object
      properties:
        fromCardId:
          type: integer
          format: int64
          description: ID карты-отправителя
          example: 123
        toCardId:
          type: integer
          format: int64
          description: ID карты-получателя
          example: 456
        amount:
          type: number
          description: Сумма перевода
          example: 100.5
          minimum: 0.01
      required:
      - fromCardId
      - toCardId
    Page:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items: {}
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    DeleteUserResponse:
      type: object
      description: Ответ после удаления пользователя
      properties:
        userID:
          type: integer
          format: int64
          description: Идентификатор удалённого пользователя
          example: 123
        message:
          type: string
          description: Сообщение о результате удаления
          example: Пользователь успешно удалён
...
  securitySchemes:
    bearerAuth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: JWT